pyproject.toml
[project]
name = "streamlit-text-summarization"
version = "0.1.0"
description = "A Streamlit app for text summarization using Google's Gemini AI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "streamlit",
    "google-generativeai",
    "python-dotenv"
]
